+++ tests/stress_tests/test_closeMeasure.py
@@ -25,6 +25,7 @@ from mw4.mainApp import MountWizzard4
 from pathlib import Path
 from PySide6.QtCore import Qt, QThreadPool
 from random import randint
+import shutil
 
 mwglob = {
     "dataDir": Path("tests/work/assets"),
@@ -45,13 +46,16 @@ def module_setup_teardown():
     tp = QThreadPool()
 
     for d in mwglob:
+        path = Path(mwglob[d])
+        path.mkdir(parents=True, exist_ok=True)
         files = glob.glob(f"{mwglob[d]}/*")
         if "modelData" in d:
             continue
         for f in files:
             if "empty" in f:
                 continue
-            os.remove(f)
+            if os.path.isfile(f):
+                os.remove(f)
 
     extractDataFiles(mwGlob=mwglob)
 
@@ -64,16 +68,14 @@ def module_setup_teardown():
         for f in files:
             if "empty" in f:
                 continue
-            os.remove(f)
+            if os.path.isfile(f):
+                os.remove(f)
 
     tp.waitForDone(3000)
 
 
 def test_1(qtbot, qapp):
-    def run():
-        qapp.exec_()
-
-    app = MountWizzard4(mwGlob=mwglob, application=qapp)
+    app = MountWizzard4(mwGlob=mwglob, application=qapp, test=1)
     count = 59
     app.measure.data = {
         "time": np.empty(shape=[0, 1], dtype="datetime64"),
@@ -87,14 +89,12 @@ def test_1(qtbot, qapp):
         value = np.datetime64(f"2014-12-12 20:20:{count:02d}")
         app.measure.data["time"] = np.append(app.measure.data["time"], value)
 
-    worker = Worker(run)
-    tp.start(worker)
     qtbot.waitExposed(app.mainW, timeout=1000)
     sleepAndEvents(100)
 
     for index in range(5):
         qtbot.mouseClick(app.mainW.ui.openMeasureW, Qt.LeftButton)
-        c = app.uiWindows["showMeasureW"]["classObj"]
+        c = app.mainW.externalWindows.uiWindows["showMeasureW"]["classObj"]
         qtbot.waitExposed(c, timeout=3000)
         c.ui.set0.setCurrentIndex(3)
         sleepAndEvents(50)
@@ -102,4 +102,5 @@ def test_1(qtbot, qapp):
         sleepAndEvents(randint(500, 1500))
         c.close()
         sleepAndEvents(50)
-    qtbot.mouseClick(app.mainW.ui.saveConfigQuit, Qt.LeftButton)
+    with qtbot.waitSignal(app.timeMgr.update10s, timeout=15000, raising=True):
+        pass
